Make improvements
