+++ scripts/python/requirements.txt
@@ -1,0 +1,19 @@
+boto3==1.14.49
+botocore==1.17.49
+certifi==2020.6.20
+chardet==3.0.4
+docutils==0.15.2
+futures==3.3.0
+gitdb2==2.0.6
+GitPython==2.1.15
+idna==2.10
+jmespath==0.10.0
+jq==1.0.2
+pathlib==1.0.1
+python-dateutil==2.8.1
+requests==2.24.0
+s3transfer==0.3.3
+six==1.15.0
+smmap==3.0.4
+smmap2==3.0.1
+urllib3==1.25.10
